refactor:#93 게시글 내 상호작용 영역 로직분리
